0.9

+++ action.demo_truncate.php
@@ -1,9 +1,10 @@
 <?php
 if(!isset($gCms)) exit;
-$retourid = 'essai';
-$cg_ops = new CGExtensions;
-$ops = new adherents_spid;
-$page = $cg_ops->resolve_alias_or_id($retourid);
+if (!$this->CheckPermission('Ping Delete'))
+{
+    echo $this->ShowErrors($this->Lang('needpermission'));
+	return;
+}
 
 $db = cmsms()->GetDb();
 $query="TRUNCATE ".cms_db_prefix()."module_ping_equipes";
@@ -27,9 +28,6 @@ $dbresult = $db->Execute($query);
 $query="TRUNCATE ".cms_db_prefix()."module_ping_sit_mens";
 $dbresult = $db->Execute($query);
 
-$query="TRUNCATE ".cms_db_prefix()."module_ping_adversaires";
-$dbresult = $db->Execute($query);
-
 $query="TRUNCATE ".cms_db_prefix()."module_ping_classement";
 $dbresult = $db->Execute($query);
 
@@ -39,12 +37,9 @@ $dbresult = $db->Execute($query);
 $query="TRUNCATE ".cms_db_prefix()."module_ping_rencontres_parties";
 $dbresult = $db->Execute($query);
 
-$query="TRUNCATE ".cms_db_prefix()."module_ping_joueurs";
+$query = "DELETE FROM ".cms_db_prefix()."module_ping_type_competitions WHERE idorga != '100001'";
 $dbresult = $db->Execute($query);
 
-$query = "DELETE FROM ".cms_db_prefix()."module_ping_type_competitions" WHERE idorga != '100001';
-$dbresult = $db->Execute($query);
-
-$this->RedirectForFrontEnd($id, $returnid, 'getInitialisation', array("step"=>"2"),  $inline='true');		
+$this->RedirectToAdminTab('compte');		
 ?>
